Replace int

+++ tiny_dnn/core/kernels/tiny_quantized_fully_connected_kernel.h
@@ -241,7 +241,7 @@ inline void tiny_quantized_fully_connected_back_kernel(
 
          if (params.has_bias_) {
            // vec_t& db = *in_grad[2];
-           for (int i = r.begin(); i < r.end(); i++) {
+           for (size_t i = r.begin(); i < r.end(); i++) {
              db[i] += curr_delta[i];
            }
          }
